chore: route admin users to /admin and update UI
