+++ mobile/package.json
@@ -44,8 +44,8 @@
   "devDependencies": {
     "@expo/ngrok": "^4.1.3",
     "@types/react": "~19.1.0",
-    "dotenv": "^17.2.3",
     "@types/react-native": "^0.72.8",
+    "dotenv": "^17.2.3",
     "eslint": "^9.25.0",
     "eslint-config-expo": "~10.0.0",
     "eslint-config-prettier": "^10.1.8",
